Fixed data remaining static
Earlier, once the API was called, the program just ran through the loop and searched the same data again and again. With the events in live matches, the program was displaying the details it caught at first attempt. By including it inside the loop, the program now calls the API during every iteration. ( This obviously takes more time than earlier )